requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.32.0
gspread>=5.12.0
google-auth>=2.28.0
aiohttp>=3.9.0